Add files via upload
This is not entirely my code, as I only changed two lines of code. The goal of this project was to add two lines of code to create a second sine wave that was 180 degress out of phase from the first sine wave. To do this I calculated a second index by adding PI to the angle in radians (Line 13), and added an OR comparison for the second index (Line 17).